Fix database health check

+++ eplmp-server/eplmp-server-ejb/src/main/java/org/polarsys/eplmp/server/PlatformHealthManagerBean.java
@@ -55,7 +55,7 @@ public class PlatformHealthManagerBean implements IPlatformHealthManagerLocal {
 
         // Database check
         try {
-            Long two = (Long) em.createNativeQuery("select 1+1").getSingleResult();
+            Integer two = (Integer) em.createNativeQuery("select 1+1").getSingleResult();
             if (two != 2) {
                 LOGGER.log(Level.SEVERE, "Database doesn't seem to be reachable");
                 check = false;
